requirements.txt.txt
streamlit
lamini-sdk
pandas
python-dotenv
